+++ bindings/angular1/directives/tab.js
@@ -9,7 +9,7 @@
     return {
       restrict: 'E',
       link: function(scope, element, attrs) {
-        var view = new GenericView(scope, element, attrs);
+        var view = GenericView.register(scope, element, attrs, {viewKey: 'ons-tab'});
         element[0].pageLoader = $onsen.createPageLoader(view);
 
         $onsen.fireComponentEvent(element[0], 'init');