+++ bindings/react/src/components/Tabbar.jsx
@@ -64,14 +64,13 @@ class Tabbar extends BasicComponent {
     node.removeEventListener('reactive', this.onReactive);
   }
 
-  componentDidUpdate(prevProps) {
-    super.componentDidUpdate(prevProps);
+  componentWillReceiveProps(nextProps) {
     const node = this._tabbar;
-    if (this.props.index !== node.getActiveTabIndex()) {
-      node.setActiveTab(this.props.index);
+    if (nextProps.index !== this.props.index && nextProps.index !== node.getActiveTabIndex()) {
+      node.setActiveTab(nextProps.index, { reject: false });
     }
-    if (this.props.onSwipe !== prevProps.onSwipe) {
-      node.onSwipe = this.props.onSwipe;
+    if (this.props.onSwipe !== nextProps.onSwipe) {
+      node.onSwipe = nextProps.onSwipe;
     }
   }
 
@@ -160,7 +159,7 @@ Tabbar.propTypes = {
    * @type object
    * @required false
    * @description
-   *  [en]Specify the animation's duration, delay and timing. E.g.  `{duration: 0.2, delay: 0.4, timing: 'ease-in'}`.[/en]
+   *  [en]Specify the animation's duration, delay and timing. E.g. `{duration: 0.2, delay: 0.4, timing: 'ease-in'}`.[/en]
    *  [ja][/ja]
    */
   animationOptions: PropTypes.object,